Capella/requirements.txt
python-dotenv
couchbase
sentence-transformers
numpy
fastapi
uvicorn
datasets
tqdm
streamlit
pandas
requests 